# contracts/openapi/portfolio-api.json
{
  "components": {
    "schemas": {
...
      "AcknowledgementRequest": {
        "properties": {
          "decisionType": {
            "pattern": "HANDLED|DEFERRED|IGNORED",
            "type": "string"
          },
          "idempotencyKey": {
            "minLength": 1,
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "reasonTags": {
            "items": {
              "enum": [
                "NEW_FUNDAMENTAL_EVIDENCE",
                "VALUATION",
                "PRICE_CONFIRMATION",
                "CATALYST",
                "RISK_REDUCTION",
                "COST_BASIS_ANCHOR",
                "HISTORICAL_HIGH_ANCHOR",
                "LOSS_AVERSION",
                "SOCIAL_IDEA"
              ],
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          }
        },
        "required": ["idempotencyKey"],
        "type": "object"
      },
      "AcknowledgementResponse": {
        "properties": {
          "acknowledgedAt": {
            "format": "date-time",
            "type": "string"
          },
          "decisionType": {
            "type": "string"
          },
          "executionSubmitted": {
            "type": "boolean"
          },
          "newlyAcknowledged": {
            "type": "boolean"
          },
          "recommendationId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PositionResponse": {
        "properties": {
          "accountId": {
            "format": "uuid",
            "type": "string"
          },
          "averageCost": {
            "type": "string"
          },
          "bucket": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "classificationConfirmed": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "marketValue": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private, manual-execution-only portfolio decision support API",
    "title": "Portfolio Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/positions/{id}": {
      "get": {
        "operationId": "position",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["portfolio-controller"]
      }
    },
...
    "/api/v1/recommendations/{id}/acknowledge": {
      "post": {
        "operationId": "acknowledge",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AcknowledgementResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["recommendation-acknowledgement-controller"]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ]
}